Adding a hack for the % scoping problem (test/shell/Test7).
The problem was caused by the following - when an implicit rule was initially
created, each shell line in the rule was turned into ValStringExp, capturing
the static environment in which the rule was created was. Later on, this
environment was used to evaluate the string value into a pipe, thus ignoring
the value for $% that came from the implicit rule instantiation.

This implements the following workaround:
 - Delays ValStringExp packing (instead the CommandValue keeps the statis env
   and the string_exp separately)
 - venv_find_implicit_rules updates all the CommandValue static envs with
   an appropriate static mapping for wild_var


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.6@13242 5109f7fa-8300-0410-942d-a7a480986ffa