Fix omake on netbsd
Omake fails to build on NetBSD, because the use of a -Werror in one
compilation step causes a warning about a kqueue-related typecast to
become an error and aborts build. The native NetBSD package building
framework, pkgsrc, contains a patch to workaround this issue. This
commit reproduces that patch here.